Refactor MahasiswaWali and MahasiswaDetail models
- Added 'kebutuhan_khusus' attribute to MahasiswaWali model
- Added 'telp_rumah' and 'kode_pos' attributes to MahasiswaDetail model

+++ routes/web.php
@@ -71,8 +71,8 @@ Route::group(['middleware' => ['auth:hr']], function () {
 
     // Bayar
     Route::post('/mahasiswa/bayar', [MahasiswaController::class, 'bayar'])->name('mahasiswa.bayar')->middleware(['permission:update_mahasiswa']);
-    Route::get('/get-paket-matakuliah-details/{id}', [MahasiswaController::class, 'getPaketMatakuliahDetails'])->middleware(['permission:update_mahasiswa']);
-    Route::get('/get-paket-matakuliah-by-semester', [MahasiswaController::class, 'getPaketMataKuliahBySemester'])->middleware(['permission:update_mahasiswa']);
+    Route::get('/get-kurikulum-details/{id}', [MahasiswaController::class, 'getKurikulumDetails'])->middleware(['permission:update_mahasiswa']);
+    Route::get('/get-kurikulum-by-semester', [MahasiswaController::class, 'getKurikulumBySemester'])->middleware(['permission:update_mahasiswa']);
 
     // ajax
     Route::get('/mahasiswa/cities/{provinceCode}', [MahasiswaController::class, 'getCities'])->middleware(['permission:read_mahasiswa']);
@@ -176,22 +176,22 @@ Route::group(['middleware' => ['auth:hr']], function () {
     Route::get('/getMahasiswaByProdi', [PrestasiController::class, 'getMahasiswaByProdi'])->name('getMahasiswaByProdi');
 
     //kuriukulum
-    Route::get('/kurikulum', [KurikulumController::class, 'index'])->name('kurikulum.index');
-    Route::get('/kurikulum/create', [KurikulumController::class, 'create'])->name('kurikulum.create');
-    Route::post('/kurikulum', [KurikulumController::class, 'store'])->name('kurikulum.store');
-    Route::get('/kurikulum/{kurikulum}/edit', [KurikulumController::class, 'edit'])->name('kurikulum.edit');
-    Route::put('/kurikulum/{kurikulum}', [KurikulumController::class, 'update'])->name('kurikulum.update');
-    Route::delete('/kurikulum/{kurikulum}', [KurikulumController::class, 'destroy'])->name('kurikulum.destroy');
-    Route::get('/kurikulum/show/{id}', [KurikulumController::class, 'show'])->name('kurikulum.show');
+    Route::get('/kurikulum', [KurikulumController::class, 'index'])->name('kurikulum.index')->middleware(['permission:read_kurikulum']);
+    Route::get('/kurikulum/create', [KurikulumController::class, 'create'])->name('kurikulum.create')->middleware(['permission:create_kurikulum']);
+    Route::post('/kurikulum', [KurikulumController::class, 'store'])->name('kurikulum.store')->middleware(['permission:create_kurikulum']);
+    Route::get('/kurikulum/{kurikulum}/edit', [KurikulumController::class, 'edit'])->name('kurikulum.edit')->middleware(['permission:update_kurikulum']);
+    Route::put('/kurikulum/{kurikulum}', [KurikulumController::class, 'update'])->name('kurikulum.update')->middleware(['permission:update_kurikulum']);
+    Route::delete('/kurikulum/{kurikulum}', [KurikulumController::class, 'destroy'])->name('kurikulum.destroy')->middleware(['permission:delete_kurikulum']);
+    Route::get('/kurikulum/show/{id}', [KurikulumController::class, 'show'])->name('kurikulum.show')->middleware(['permission:read_kurikulum']);
 
     //kelas
-    Route::get('/kelas', [KelasController::class, 'index'])->name('kelast.index');
-    Route::get('/kelas/create', [KelasController::class, 'create'])->name('kelast.create');
-    Route::post('/kelas', [KelasController::class, 'store'])->name('kelast.store');
-    Route::get('/kelas/{kelas}/edit', [KelasController::class, 'edit'])->name('kelast.edit');
-    Route::put('/kelas/{kelas}', [KelasController::class, 'update'])->name('kelast.update');
-    Route::delete('/kelas/{kelas}', [KelasController::class, 'destroy'])->name('kelast.destroy');
-    Route::get('/kelas/show/{id}', [KelasController::class, 'show'])->name('kelast.show');
+    Route::get('/kelas', [KelasController::class, 'index'])->name('kelast.index')->middleware(['permission:read_kelas']);
+    Route::get('/kelas/create', [KelasController::class, 'create'])->name('kelast.create')->middleware(['permission:create_kelas']);
+    Route::post('/kelas', [KelasController::class, 'store'])->name('kelast.store')->middleware(['permission:create_kelas']);
+    Route::get('/kelas/{kelas}/edit', [KelasController::class, 'edit'])->name('kelast.edit')->middleware(['permission:update_kelas']);
+    Route::put('/kelas/{kelas}', [KelasController::class, 'update'])->name('kelast.update')->middleware(['permission:update_kelas']);
+    Route::delete('/kelas/{kelas}', [KelasController::class, 'destroy'])->name('kelast.destroy')->middleware(['permission:delete_kelas']);
+    Route::get('/kelas/show/{id}', [KelasController::class, 'show'])->name('kelast.show')->middleware(['permission:read_kelas']);
     Route::get('/kelas/details/{kurikulum}', [KelasController::class, 'getKurikulumDetails'])->name('kelast.details');
 
     // Route untuk mengambil mata kuliah berdasarkan program studi dan semester
